scene: Add character control UI
Add character control debug UI under the common ui_debug interface so that
it can be switched on/off from one place.

+++ core/character.h
@@ -51,7 +51,7 @@ struct character {
     vec3    motion;
     vec3    angle;
     vec3    normal;
-    double  speed;
+    float   speed;
     float   yaw_turn;
     float   pitch_turn;
     struct list entry;
